TEIID-3393: separate AdminFactory and AdminImpl from EmbeddedServer
